Flash HAL driver implemented.

+++ trunk/Src/HAL/Flash/base/Flash_api.c
@@ -141,7 +141,7 @@ __weak eError onReadHregFlashData( uint8_t instance, const HREG* hregPtr, void* 
 
 	uint32_t offset;
 	READ_HREG( instance, HREG_FLASH_OFFSET, &offset);
-	result = FlashReadData((tFlash)instance, (uint16_t*) param1, offset);
+	result = FlashReadData((tFlash)instance, (uint32_t*) param1, offset);
 
 	return result;
 }
@@ -156,7 +156,7 @@ __weak eError onWriteHregFlashData( uint8_t instance, const HREG* hregPtr, uint3
 
 	uint32_t offset;
 	READ_HREG( instance, HREG_FLASH_OFFSET, &offset);
-	result = FlashProgramData((tFlash)instance, (uint16_t) param1, offset);
+	result = FlashProgramData((tFlash)instance, (uint64_t) param1, offset);
 
 	return result;
 }
